requirements.txt
python==3.6.3
pytorch==1.2.0
torchtext==0.2.1
numpy==1.16.2
scipy==1.2.1